+++ src/logic/wrappers/withExtendableCacheAsync.test.ts
@@ -1,25 +1,25 @@
-import { createExampleSyncCache } from '../../__test_assets__/createExampleCache';
-import { SimpleCache } from '../../domain/SimpleCache';
+import { createExampleAsyncCache } from '../../__test_assets__/createExampleCache';
+import { SimpleAsyncCache } from '../../domain/SimpleCache';
 import { BadRequestError } from '../../utils/errors/BadRequestError';
-import { withExtendableCaching } from './withExtendableCaching';
+import { withExtendableCacheAsync } from './withExtendableCacheAsync';
 
-describe('withExtendableCaching', () => {
+describe('withExtendableCacheAsync', () => {
   describe('execute', () => {
-    it('should be able to execute logic', () => {
+    it('should be able to execute logic', async () => {
       // define an example fn
       const apiCalls = [];
-      const callApi = withExtendableCaching(
-        () => {
+      const callApi = withExtendableCacheAsync(
+        async () => {
           apiCalls.push(1);
           return Math.random();
         },
-        { cache: createExampleSyncCache().cache },
+        { cache: createExampleAsyncCache().cache },
       );
 
       // call the fn a few times
-      const result1 = callApi.execute();
-      const result2 = callApi.execute();
-      const result3 = callApi.execute();
+      const result1 = await callApi.execute();
+      const result2 = await callApi.execute();
+      const result3 = await callApi.execute();
 
       // check that the response is the same each time
       expect(result1).toEqual(result2);
@@ -30,22 +30,22 @@ describe('withExtendableCaching', () => {
     });
   });
   describe('invalidate', () => {
-    it('should be able to invalidate a cached value by input', () => {
+    it('should be able to invalidate a cached value by input', async () => {
       // define an example fn
       const apiCalls = [];
-      const callApi = withExtendableCaching(
-        ({ galaxy }: { galaxy: string }) => {
+      const callApi = withExtendableCacheAsync(
+        async ({ galaxy }: { galaxy: string }) => {
           apiCalls.push(galaxy);
           return Math.random();
         },
-        { cache: createExampleSyncCache().cache },
+        { cache: createExampleAsyncCache().cache },
       );
 
       // call the fn a few times
-      const result1 = callApi.execute({ galaxy: 'andromeda' });
-      const result2 = callApi.execute({ galaxy: 'pegasus' });
-      const result3 = callApi.execute({ galaxy: 'andromeda' });
-      const result4 = callApi.execute({ galaxy: 'pegasus' });
+      const result1 = await callApi.execute({ galaxy: 'andromeda' });
+      const result2 = await callApi.execute({ galaxy: 'pegasus' });
+      const result3 = await callApi.execute({ galaxy: 'andromeda' });
+      const result4 = await callApi.execute({ galaxy: 'pegasus' });
 
       // check that the response is the same each time the input is the same
       expect(result1).toEqual(result3);
@@ -55,29 +55,29 @@ describe('withExtendableCaching', () => {
       expect(apiCalls.length).toEqual(2);
 
       // invalidate the cached value for one of the inputs
-      callApi.invalidate({ forInput: [{ galaxy: 'andromeda' }] });
+      await callApi.invalidate({ forInput: [{ galaxy: 'andromeda' }] });
 
       // now call the cache again for that invalidated value, and prove it called the api again
-      const result5 = callApi.execute({ galaxy: 'andromeda' });
+      const result5 = await callApi.execute({ galaxy: 'andromeda' });
       expect(result5).not.toEqual(result1);
       expect(apiCalls.length).toEqual(3);
     });
-    it('should be able to invalidate a cached value by key', () => {
+    it('should be able to invalidate a cached value by key', async () => {
       // define an example fn
       const apiCalls = [];
-      const callApi = withExtendableCaching(
-        ({ galaxy }: { galaxy: string }) => {
+      const callApi = withExtendableCacheAsync(
+        async ({ galaxy }: { galaxy: string }) => {
           apiCalls.push(galaxy);
           return Math.random();
         },
-        { cache: createExampleSyncCache().cache },
+        { cache: createExampleAsyncCache().cache },
       );
 
       // call the fn a few times
-      const result1 = callApi.execute({ galaxy: 'andromeda' });
-      const result2 = callApi.execute({ galaxy: 'pegasus' });
-      const result3 = callApi.execute({ galaxy: 'andromeda' });
-      const result4 = callApi.execute({ galaxy: 'pegasus' });
+      const result1 = await callApi.execute({ galaxy: 'andromeda' });
+      const result2 = await callApi.execute({ galaxy: 'pegasus' });
+      const result3 = await callApi.execute({ galaxy: 'andromeda' });
+      const result4 = await callApi.execute({ galaxy: 'pegasus' });
 
       // check that the response is the same each time the input is the same
       expect(result1).toEqual(result3);
@@ -87,19 +87,24 @@ describe('withExtendableCaching', () => {
       expect(apiCalls.length).toEqual(2);
 
       // invalidate the cached value for one of the inputs
-      callApi.invalidate({ forKey: JSON.stringify([{ galaxy: 'andromeda' }]) });
+      await callApi.invalidate({
+        forKey: JSON.stringify([{ galaxy: 'andromeda' }]),
+      });
 
       // now call the cache again for that invalidated value, and prove it called the api again
-      const result5 = callApi.execute({ galaxy: 'andromeda' });
+      const result5 = await callApi.execute({ galaxy: 'andromeda' });
       expect(result5).not.toEqual(result1);
       expect(apiCalls.length).toEqual(3);
     });
-    it('should be able to invalidate a cached value by key when the cache is to be defined at runtime from inputs', () => {
+    it('should be able to invalidate a cached value by key when the cache is to be defined at runtime from inputs', async () => {
       // define an example fn
-      const { cache } = createExampleSyncCache();
+      const { cache } = createExampleAsyncCache<string>();
       const apiCalls = [];
-      const callApi = withExtendableCaching(
-        ({ galaxy }: { galaxy: string }, _: { cache: SimpleCache<string> }) => {
+      const callApi = withExtendableCacheAsync(
+        async (
+          { galaxy }: { galaxy: string },
+          _: { cache: SimpleAsyncCache<string> },
+        ) => {
           apiCalls.push(galaxy);
           return Math.random();
         },
@@ -112,10 +117,10 @@ describe('withExtendableCaching', () => {
       );
 
       // call the fn a few times
-      const result1 = callApi.execute({ galaxy: 'andromeda' }, { cache });
-      const result2 = callApi.execute({ galaxy: 'pegasus' }, { cache });
-      const result3 = callApi.execute({ galaxy: 'andromeda' }, { cache });
-      const result4 = callApi.execute({ galaxy: 'pegasus' }, { cache });
+      const result1 = await callApi.execute({ galaxy: 'andromeda' }, { cache });
+      const result2 = await callApi.execute({ galaxy: 'pegasus' }, { cache });
+      const result3 = await callApi.execute({ galaxy: 'andromeda' }, { cache });
+      const result4 = await callApi.execute({ galaxy: 'pegasus' }, { cache });
 
       // check that the response is the same each time the input is the same
       expect(result1).toEqual(result3);
@@ -126,7 +131,7 @@ describe('withExtendableCaching', () => {
 
       // prove that it will throw a helpful error if we dont explicitly pass in the cache in this case
       try {
-        callApi.invalidate({ forKey: 'andromeda' });
+        await callApi.invalidate({ forKey: 'andromeda' });
         throw new Error('should not reach here');
       } catch (error) {
         expect(error).toBeInstanceOf(BadRequestError);
@@ -139,31 +144,31 @@ describe('withExtendableCaching', () => {
       }
 
       // invalidate the cached value for one of the inputs
-      callApi.invalidate({ forKey: 'andromeda', cache });
+      await callApi.invalidate({ forKey: 'andromeda', cache });
 
       // now call the cache again for that invalidated value, and prove it called the api again
-      const result5 = callApi.execute({ galaxy: 'andromeda' }, { cache });
+      const result5 = await callApi.execute({ galaxy: 'andromeda' }, { cache });
       expect(result5).not.toEqual(result1);
       expect(apiCalls.length).toEqual(3);
     });
   });
   describe('update', () => {
-    it('should be able to update a cached value by input', () => {
+    it('should be able to update a cached value by input', async () => {
       // define an example fn
       const apiCalls = [];
-      const callApi = withExtendableCaching(
-        ({ galaxy }: { galaxy: string }) => {
+      const callApi = withExtendableCacheAsync(
+        async ({ galaxy }: { galaxy: string }) => {
           apiCalls.push(galaxy);
           return Math.random();
         },
-        { cache: createExampleSyncCache().cache },
+        { cache: createExampleAsyncCache().cache },
       );
 
       // call the fn a few times
-      const result1 = callApi.execute({ galaxy: 'andromeda' });
-      const result2 = callApi.execute({ galaxy: 'pegasus' });
-      const result3 = callApi.execute({ galaxy: 'andromeda' });
-      const result4 = callApi.execute({ galaxy: 'pegasus' });
+      const result1 = await callApi.execute({ galaxy: 'andromeda' });
+      const result2 = await callApi.execute({ galaxy: 'pegasus' });
+      const result3 = await callApi.execute({ galaxy: 'andromeda' });
+      const result4 = await callApi.execute({ galaxy: 'pegasus' });
 
       // check that the response is the same each time the input is the same
       expect(result1).toEqual(result3);
@@ -173,10 +178,13 @@ describe('withExtendableCaching', () => {
       expect(apiCalls.length).toEqual(2);
 
       // update the cached value for one of the inputs
-      callApi.update({ forInput: [{ galaxy: 'andromeda' }], toValue: 821 });
+      await callApi.update({
+        forInput: [{ galaxy: 'andromeda' }],
+        toValue: 821,
+      });
 
       // now call the cache again
-      const result5 = callApi.execute({ galaxy: 'andromeda' });
+      const result5 = await callApi.execute({ galaxy: 'andromeda' });
 
       // and prove that the value changed to what we wanted to update it to
       expect(result5).toEqual(821);
@@ -185,22 +193,22 @@ describe('withExtendableCaching', () => {
       // and prove that we didn't call the api again
       expect(apiCalls.length).toEqual(2);
     });
-    it('should be able to update a cached value by key', () => {
+    it('should be able to update a cached value by key', async () => {
       // define an example fn
       const apiCalls = [];
-      const callApi = withExtendableCaching(
-        ({ galaxy }: { galaxy: string }) => {
+      const callApi = withExtendableCacheAsync(
+        async ({ galaxy }: { galaxy: string }) => {
           apiCalls.push(galaxy);
           return Math.random();
         },
-        { cache: createExampleSyncCache().cache },
+        { cache: createExampleAsyncCache().cache },
       );
 
       // call the fn a few times
-      const result1 = callApi.execute({ galaxy: 'andromeda' });
-      const result2 = callApi.execute({ galaxy: 'pegasus' });
-      const result3 = callApi.execute({ galaxy: 'andromeda' });
-      const result4 = callApi.execute({ galaxy: 'pegasus' });
+      const result1 = await callApi.execute({ galaxy: 'andromeda' });
+      const result2 = await callApi.execute({ galaxy: 'pegasus' });
+      const result3 = await callApi.execute({ galaxy: 'andromeda' });
+      const result4 = await callApi.execute({ galaxy: 'pegasus' });
 
       // check that the response is the same each time the input is the same
       expect(result1).toEqual(result3);
@@ -210,13 +218,13 @@ describe('withExtendableCaching', () => {
       expect(apiCalls.length).toEqual(2);
 
       // update the cached value for one of the inputs
-      callApi.update({
+      await callApi.update({
         forKey: JSON.stringify([{ galaxy: 'andromeda' }]),
         toValue: 821,
       });
 
       // now call the cache again
-      const result5 = callApi.execute({ galaxy: 'andromeda' });
+      const result5 = await callApi.execute({ galaxy: 'andromeda' });
 
       // and prove that the value changed to what we wanted to update it to
       expect(result5).toEqual(821);
@@ -225,12 +233,15 @@ describe('withExtendableCaching', () => {
       // and prove that we didn't call the api again
       expect(apiCalls.length).toEqual(2);
     });
-    it('should be able to update a cached value by key when the cache is to be defined at runtime from inputs', () => {
+    it('should be able to update a cached value by key when the cache is to be defined at runtime from inputs', async () => {
       // define an example fn
-      const { cache } = createExampleSyncCache();
+      const { cache } = createExampleAsyncCache<string>();
       const apiCalls = [];
-      const callApi = withExtendableCaching(
-        ({ galaxy }: { galaxy: string }, _: { cache: SimpleCache<string> }) => {
+      const callApi = withExtendableCacheAsync(
+        async (
+          { galaxy }: { galaxy: string },
+          _: { cache: SimpleAsyncCache<string> },
+        ) => {
           apiCalls.push(galaxy);
           return Math.random();
         },
@@ -243,10 +254,10 @@ describe('withExtendableCaching', () => {
       );
 
       // call the fn a few times
-      const result1 = callApi.execute({ galaxy: 'andromeda' }, { cache });
-      const result2 = callApi.execute({ galaxy: 'pegasus' }, { cache });
-      const result3 = callApi.execute({ galaxy: 'andromeda' }, { cache });
-      const result4 = callApi.execute({ galaxy: 'pegasus' }, { cache });
+      const result1 = await callApi.execute({ galaxy: 'andromeda' }, { cache });
+      const result2 = await callApi.execute({ galaxy: 'pegasus' }, { cache });
+      const result3 = await callApi.execute({ galaxy: 'andromeda' }, { cache });
+      const result4 = await callApi.execute({ galaxy: 'pegasus' }, { cache });
 
       // check that the response is the same each time the input is the same
       expect(result1).toEqual(result3);
@@ -257,7 +268,7 @@ describe('withExtendableCaching', () => {
 
       // prove that it will throw a helpful error if we dont explicitly pass in the cache in this case
       try {
-        callApi.update({ forKey: 'andromeda', toValue: 821 });
+        await callApi.update({ forKey: 'andromeda', toValue: 821 });
         throw new Error('should not reach here');
       } catch (error) {
         expect(error).toBeInstanceOf(BadRequestError);
@@ -268,10 +279,10 @@ describe('withExtendableCaching', () => {
       }
 
       // invalidate the cached value for one of the inputs
-      callApi.update({ forKey: 'andromeda', toValue: 821, cache });
+      await callApi.update({ forKey: 'andromeda', toValue: 821, cache });
 
       // now call the cache again
-      const result5 = callApi.execute({ galaxy: 'andromeda' }, { cache });
+      const result5 = await callApi.execute({ galaxy: 'andromeda' }, { cache });
 
       // and prove that the value changed to what we wanted to update it to
       expect(result5).toEqual(821);
@@ -280,22 +291,22 @@ describe('withExtendableCaching', () => {
       // and prove that we didn't call the api again
       expect(apiCalls.length).toEqual(2);
     });
-    it('should be able to update a cached value by input to a value based on prior value', () => {
+    it('should be able to update a cached value by input to a value based on prior value', async () => {
       // define an example fn
       const apiCalls = [];
-      const callApi = withExtendableCaching(
-        ({ galaxy }: { galaxy: string }) => {
+      const callApi = withExtendableCacheAsync(
+        async ({ galaxy }: { galaxy: string }) => {
           apiCalls.push(galaxy);
           return Math.random();
         },
-        { cache: createExampleSyncCache().cache },
+        { cache: createExampleAsyncCache().cache },
       );
 
       // call the fn a few times
-      const result1 = callApi.execute({ galaxy: 'andromeda' });
-      const result2 = callApi.execute({ galaxy: 'pegasus' });
-      const result3 = callApi.execute({ galaxy: 'andromeda' });
-      const result4 = callApi.execute({ galaxy: 'pegasus' });
+      const result1 = await callApi.execute({ galaxy: 'andromeda' });
+      const result2 = await callApi.execute({ galaxy: 'pegasus' });
+      const result3 = await callApi.execute({ galaxy: 'andromeda' });
+      const result4 = await callApi.execute({ galaxy: 'pegasus' });
 
       // check that the response is the same each time the input is the same
       expect(result1).toEqual(result3);
@@ -305,13 +316,13 @@ describe('withExtendableCaching', () => {
       expect(apiCalls.length).toEqual(2);
 
       // update the cached value for one of the inputs
-      callApi.update({
+      await callApi.update({
         forInput: [{ galaxy: 'andromeda' }],
         toValue: ({ fromCachedOutput }) => (fromCachedOutput ?? 0) * 2,
       });
 
       // now call the cache again
-      const result5 = callApi.execute({ galaxy: 'andromeda' });
+      const result5 = await callApi.execute({ galaxy: 'andromeda' });
 
       // and prove that the value changed to what we wanted to update it to
       expect(result5).not.toEqual(result1);